Fold the FRI challenge powers and evals into per-proof constants
computeFRIExpression combined the DEEP quotient with two Horner chains:

  accum_o = SUM_j vf2^(n_o-1-j) * (p_j[r] - e_j)
  fri     = SUM_o vf1^(O-1-o) * inv_g[o] * accum_o

Every vf1/vf2 power and every opening evaluation e_j is a per-proof
constant, not a per-row value, so both chains collapse to

  coef_(o,j) = vf1^(O-1-o) * vf2^(n_o-1-j)   K_o = -SUM_j coef_(o,j)*e_j
  fri = SUM_o inv_g[o] * ( SUM_j coef_j * p_j[r] + K_o )

with the constants computed once per proof by computeFRIFoldedConstants.
The row loop becomes one mul31 + one add33 per base-field column instead
of a sub33 + mul33 + add33 -- half the muls and half the adds -- the
evals load leaves the row loop entirely, and the outer multiply by vf1
disappears since those powers fold into coef. inv_g[o] is row-dependent,
so the per-opening grouping and its batched inversion are unchanged, and
an opening with no columns needs no special case (K_o = 0).

Measured on an RTX 5090 at the six zisk AIR shapes (openingPoints, evMap
and distinct eval-map columns taken from their starkinfo), 3 runs:

  Keccakf +9.9%  Sha256f +36.4%  Poseidon +27.5%
  compressor +32.9%  recursive2 +29.4%  Main +36.7%

Below the 48% arithmetic saving because the kernel is memory-leaning.

The kernels move to a self-contained fri_expression.cuh so they can be
checked against a host reference without the StarkInfo translation unit,
which is why EvalInfo moves to its own header. The coefficients live in a
new per-stream aux_trace region rather than in AirInstanceInfo, which is
shared per AIR and would let concurrent proofs of one AIR race.

The verifier is untouched -- it still evaluates the friExp expression
form -- so the end-to-end proofs below were accepted by an unmodified,
independent verifier.

Tests: the host reference was validated bit-for-bit against the Horner
kernel before it was removed. Covers opening counts 1..97, exact/partial/
single inversion groups, empty openings, mixed dim 1/3, all three source
buffers, and columns shared across openings. testsgpu 45/45; prove and
prove -a on fibonacci-square verify, including the full recursion chain.

A column-grouped variant was also built and benchmarked: it removed up to
58% of the column loads and was still slower than this on 5 of 6 shapes
(occupancy-bound), so it is not included.
