+++ pil2-stark/src/starkpil/stark_info.hpp
@@ -9,23 +9,13 @@
 #include "goldilocks_base_field.hpp"
 #include "zklog.hpp"
 #include "exit_process.hpp"
+#include "eval_info.hpp"
 
 using json = nlohmann::json;
 using namespace std;
 
 /* StarkInfo class contains the contents of the file zkevm.starkinfo.json,
    which is parsed during the constructor */
-
-struct EvalInfo
-{
-    uint64_t type; // 0: cm, 1: custom, 2: fixed
-    uint64_t offset;
-    uint64_t stagePos;
-    uint64_t stageCols;
-    uint64_t dim;
-    uint64_t openingPos;
-    uint64_t evalPos;
-};
 
 typedef enum
 {
